+++ flutter_wallet/lib/screens/home_screen.dart
@@ -1,6 +1,6 @@
 import '../utils/log.dart';
+import '../utils/secure_clipboard.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 import 'package:intl/intl.dart';
 import '../services/wallet_service.dart';
@@ -121,7 +121,7 @@ class _HomeScreenState extends State<HomeScreen> {
 
   void _copyAddress() {
     if (_address != null) {
-      Clipboard.setData(ClipboardData(text: _address!));
+      SecureClipboard.copyPublic(_address!);
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(content: Text('Address copied to clipboard')),
       );